+++ lib/auto_html.rb
@@ -1,17 +1,19 @@
-%w(base filter builder auto_html_for).each do |f|
-  require File.expand_path("../auto_html/#{f}", __FILE__)
-end
+module AutoHtml
 
-Dir["#{File.dirname(__FILE__) + '/auto_html/filters'}/**/*"].each do |filter|
-  require "#{filter}"
-end
+  autoload :Filter,       'auto_html/filter'
+  autoload :Format,       'auto_html/format'
 
-# if rails
-require 'auto_html/railtie' if defined?(Rails::Railtie)
-if defined?(ActiveRecord::Base)
-  ActiveRecord::Base.send :include, AutoHtmlFor
+  autoload :Dailymotion,  'auto_html/filters/dailymotion'
+  autoload :Gist,         'auto_html/filters/gist'
+  autoload :GoogleMap,    'auto_html/filters/google_map'
+  autoload :HtmlEscape,   'auto_html/filters/html_escape'
+  autoload :Image,        'auto_html/filters/image'
+  autoload :Instagram,    'auto_html/filters/instagram'
+  autoload :Link,         'auto_html/filters/link'
+  autoload :Metacafe,     'auto_html/filters/metacafe'
+  autoload :SimpleFormat, 'auto_html/filters/simple_format'
+  autoload :Ted,          'auto_html/filters/ted'
+  autoload :Vimeo,        'auto_html/filters/vimeo'
+  autoload :Youtube,      'auto_html/filters/youtube'
 
-  module ActionView::Helpers::TextHelper
-    include AutoHtml
-  end
 end