+++ test/unit/filters/instagram_test.rb
@@ -1,8 +1,9 @@
 require File.expand_path('../../unit_test_helper', __FILE__)
 
-class InstagramTest < Minitest::Test
+class InstagramTest < Test::Unit::TestCase
   def test_instagram_embed
-    result = auto_html('http://instagram.com/p/WsQTLAGvx7/') { instagram }
+    filter = AutoHtml::Instagram.new
+    result = filter.call('http://instagram.com/p/WsQTLAGvx7/')
     assert_equal '<iframe src="http://instagram.com/p/WsQTLAGvx7/embed" height="714" width="616" frameborder="0" scrolling="no"></iframe>', result
   end
 end